Restructure and start we cleaner slate
