--- src/Algebra/SemiGroups.hs
@@ -1,23 +1,0 @@
-module Algebra.SemiGroups where
-
-{-|
- given a set S
-
-> An operation ⊗ : S x S ⟶ S (... so a binary operation)
-> If ∀ a,b ∈ S, (a ⊗ b) ∈ S (S is closed on ⊗)
-> And ∀ a,b,c ∈ S, (a ⊗ b) ⊗ c = a ⊗ (b ⊗ c) = a ⊗ b ⊗ c
-
- The structure (S, ⊗) is a semi-group.
--}
-class SemiGroup a where
-  (|+|) :: a -> a -> a
-  {-# MINIMAL (|+|) #-}
-
-instance SemiGroup [a] where
-  (|+|) = (++)
-  {-# INLINE (|+|) #-}
-
-
-instance SemiGroup Int where
-  (|+|) = (+)
-  {-# INLINE (|+|) #-}
